+++ src/constants.ts
@@ -1,1 +1,3 @@
 export const CONFIG_FILE_NAME = 'pint.json';
+export const DEFAULT_EXEC_PATH = ['vendor', 'bin', 'pint'];
+export const DEFAULT_LARAVEL_SAIL_EXEC_PATH = ['vendor', 'bin', 'sail'];